Added rudimentary error checking to file loader, and a readme.

+++ src/pyquest/__main__.py
@@ -20,18 +20,40 @@ from zipfile import ZipFile
 
 from pyquest.game import QuestGame
 
-file_name = sys.argv[1]
+VERSION = "0.1-SNAPSHOT"
 
+print("pyQuest", VERSION)
+if len(sys.argv) != 2:
+    print("Usage:")
+    print("py" if os.name == 'nt' else "python3", "-m pyquest [FILENAME]")
+    print("Where [FILENAME] is the name of a Quest game file")
+    print("(either a packed .quest file, or a raw .aslx source file)")
+    exit(1)
+
+# This will become necessary for multimedia functions.
+# try:
+#     import pygame
+# except ImportError:
+#     print("*** Module pygame not found! Pygame is required to run pyQuest.", file=sys.stderr)
+#     exit(1)
+
+file_name = sys.argv[1]
 if file_name[-4:] == ".zip" or file_name[-6:] == ".quest":
     print("Opening QUEST file...")
-    zipped = ZipFile(file_name, 'r')
+    try:
+        zipped = ZipFile(file_name, 'r')
+    except FileNotFoundError:
+        print("*** Game File", file_name, "not found!", file=sys.stderr)
+        exit(127)
     game_folder = tempfile.mkdtemp()
     zipped.extractall(game_folder)
     zipped.close()
     print(game_folder)
-    os.system('ls -lah ' + game_folder)
-    the_game = QuestGame(game_folder, from_qfile=True)
+    # os.system('ls -lah ' + game_folder)
+    the_game = QuestGame(game_folder, launch_dir=os.path.sep.join(os.path.split(file_name)[-1]),
+                         from_qfile=True)
     the_game.run()
     os.system('rm -rf ' + game_folder)
 else:
     the_game = QuestGame(sys.argv[1], from_qfile=False)
+    the_game.run()
